+++ core/src/main/java/org/nervos/ckb/methods/type/TxTrace.java
@@ -1,9 +1,0 @@
-package org.nervos.ckb.methods.type;
-
-/** Created by duanyytop on 2019-03-01. Copyright © 2018 Nervos Foundation. All rights reserved. */
-public class TxTrace {
-
-  public String action;
-  public String info;
-  public long time;
-}